+++ ksoftapi/models.py
@@ -13,3 +13,15 @@ class BanInfo:
         self.appeal_date = kwargs.get('appeal_date')
         self.requested_by = kwargs.get('requested_by')
         self.exists = kwargs.get('exists')
+
+
+class PaginatorListing:
+    def __init__(self, **kwargs):
+        self.count = kwargs.get('ban_count')
+        self.page_count = kwargs.get('page_count')
+        self.per_page = kwargs.get('per_page')
+        self.page = kwargs.get('page')
+        self.on_page = kwargs.get('on_page')
+        self.next_page = kwargs.get('next_page')
+        self.previous_page = kwargs.get('previous_page')
+        self.data = [BanInfo(**ban) for ban in kwargs.get('data')]
